Fix overflow in main execution file
The varable 'simulation' and 'echo' both are bool datatype, but their
address casted as int * as 'add_param' function pramater. The size of
bool is 1, but int is 4, this kind of misuse cause overflow.
The problem fix by adding more pramater for bool type, adding struct
member 'switches' and set the unuse paramater 'valp' or 'switches' to
NULL.